修复了可能存在的爆int的问题

+++ src/BookSystem.h
@@ -31,7 +31,7 @@ private:
   char Author[61];
   char Keyword[61];
 
-  int Quantity;
+  long long Quantity;
   double Price;
 
 };
